requirements.txt
ipykernel
ipython
jupyterlab
load-dotenv
openai
pandas
